# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "description": "Get basic information about the API and available endpoints",
        "operationId": "AppController_getInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "endpoints": {
                      "type": "object"
                    },
                    "documentation": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "API information",
        "tags": [
          "General"
        ]
      }
    },
    "/api/inventory": {
      "get": {
        "description": "Retrieve a complete list of all inventory items",
        "operationId": "InventoryController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved inventory items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 3
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InventoryItemEntity"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List all inventory items",
        "tags": [
          "Inventory"
        ]
      },
      "post": {
        "description": "Add a new item to the inventory",
        "operationId": "InventoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inventory item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Inventory item created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/InventoryItemEntity"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new inventory item",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/inventory/search": {
      "get": {
        "description": "Search for inventory items by name or description",
        "operationId": "InventoryController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search term to match against item name or description",
            "schema": {
              "example": "laptop",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 1
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InventoryItemEntity"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing search query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Search query parameter \"q\" is required"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Search inventory items",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/inventory/{id}": {
      "get": {
        "description": "Retrieve details of a single inventory item by its ID",
        "operationId": "InventoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the inventory item",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved inventory item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/InventoryItemEntity"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Inventory item not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Inventory item with ID 1 not found"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get a specific inventory item",
        "tags": [
          "Inventory"
        ]
      },
      "put": {
        "description": "Update all fields of an existing inventory item",
        "operationId": "InventoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the inventory item",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventory item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Inventory item updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/InventoryItemEntity"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Inventory item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update an existing inventory item",
        "tags": [
          "Inventory"
        ]
      },
      "delete": {
        "description": "Remove an inventory item from the system",
        "operationId": "InventoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the inventory item",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Inventory item deleted successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/InventoryItemEntity"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Inventory item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete an inventory item",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/feedback": {
      "get": {
        "description": "Retrieve a complete list of all customer feedback items with date, title, email, and description",
        "operationId": "FeedbackController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved feedback items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 3
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FeedbackItemEntity"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List all customer feedback",
        "tags": [
          "Feedback"
        ]
      },
      "post": {
        "description": "Create a new feedback item",
        "operationId": "FeedbackController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Feedback submitted successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/FeedbackItemEntity"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "summary": "Submit new customer feedback",
        "tags": [
          "Feedback"
        ]
      }
    },
    "/api/feedback/{id}": {
      "get": {
        "description": "Retrieve details of a single feedback item by its ID",
        "operationId": "FeedbackController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the feedback item",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved feedback item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/FeedbackItemEntity"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Feedback item not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Feedback item with ID 1 not found"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get a specific feedback item",
        "tags": [
          "Feedback"
        ]
      }
    }
  },
  "info": {
    "title": "Inventory Management API",
    "description": "A RESTful API for inventory management with CRUD operations. This API provides endpoints to create, read, update, and delete inventory items, as well as search functionality.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
      "InventoryItemEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Auto-generated unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the inventory item",
            "example": "Laptop"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the item",
            "example": "High-performance laptop for business use"
          },
          "quantity": {
            "type": "number",
            "description": "Available quantity",
            "example": 15
          },
          "price": {
            "type": "number",
            "description": "Price per unit",
            "example": 1299.99
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "quantity",
          "price"
        ]
      },
      "CreateInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the inventory item",
            "example": "Laptop"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the item",
            "example": "High-performance laptop for business use"
          },
          "quantity": {
            "type": "number",
            "description": "Available quantity",
            "example": 15,
            "minimum": 0
          },
          "price": {
            "type": "number",
            "description": "Price per unit",
            "example": 1299.99,
            "minimum": 0
          }
        },
        "required": [
          "name",
          "description",
          "quantity",
          "price"
        ]
      },
      "UpdateInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the inventory item",
            "example": "Updated Laptop"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the item",
            "example": "Updated description"
          },
          "quantity": {
            "type": "number",
            "description": "Available quantity",
            "example": 20,
            "minimum": 0
          },
          "price": {
            "type": "number",
            "description": "Price per unit",
            "example": 1499.99,
            "minimum": 0
          }
        },
        "required": [
          "name",
          "description",
          "quantity",
          "price"
        ]
      },
      "FeedbackItemEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Auto-generated unique identifier",
            "example": 1
          },
          "date": {
            "type": "string",
            "description": "Date of the feedback",
            "example": "2025-11-14"
          },
          "title": {
            "type": "string",
            "description": "Title of the feedback",
            "example": "Great product quality"
          },
          "email": {
            "type": "string",
            "description": "Email of the reporter",
            "example": "customer@example.com"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the feedback",
            "example": "The product exceeded my expectations. Very satisfied with the purchase."
          }
        },
        "required": [
          "id",
          "date",
          "title",
          "email",
          "description"
        ]
      },
      "CreateFeedbackItemDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Date of the feedback",
            "example": "2025-11-14"
          },
          "title": {
            "type": "string",
            "description": "Title of the feedback",
            "example": "Great product quality"
          },
          "email": {
            "type": "string",
            "description": "Email of the reporter",
            "example": "customer@example.com"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the feedback",
            "example": "The product exceeded my expectations. Very satisfied with the purchase."
          }
        },
        "required": [
          "date",
          "title",
          "email",
          "description"
        ]
      }
    }
  }
}